+++ app/index.jsx
@@ -1,0 +1,23 @@
+import { StatusBar } from 'expo-status-bar';
+import { StyleSheet, Text, View } from 'react-native-web';
+import { Link } from 'expo-router';
+
+
+export default function App() {
+  return (
+    <View style={styles.container}>
+      <Text>Aora!</Text>
+      <StatusBar style="auto" />
+      <Link href="/profile" style={{ color: 'blue' }}>Go to Profile</Link>
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+    container: {
+        flex: 1,
+        backgroundColor: '#fff',
+        alignItems: 'center',
+        justifyContent: 'center'
+    },
+});
