+++ app/profile.jsx
@@ -1,0 +1,14 @@
+import { StyleSheet, Text, View } from 'react-native'
+import React from 'react'
+
+const profile = () => {
+  return (
+    <View>
+      <Text>profile</Text>
+    </View>
+  )
+}
+
+export default profile
+
+const styles = StyleSheet.create({})
